Revisit metadata tests
